backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
redis==5.0.1
typing-extensions>=4.8.0
